requirements.txt
discord.py
requests
bs4
urllib3
re